create PaymentsOverviewTab
