+++ packages/atlas/src/views/playground/Playgrounds/index.ts
@@ -9,4 +9,3 @@ export * from './PlaygroundReactionsComments'
 export * from './PlaygroundIframe'
 export * from './PlaygroundCaptcha'
 export * from './PlaygroundGoogleAuthentication'
-export * from './PlaygroundChannelPayouts'
