+++ packages/atlas/src/views/studio/MyPaymentsView/PaymentsOverviewTab/index.ts
@@ -1,0 +1,1 @@
+export * from './PaymentsOverviewTab'
